+++ src/shared/hook/useCategoryListQuery.ts
@@ -1,6 +1,6 @@
 import { UseQueryResult, useQuery } from '@tanstack/react-query';
 
-import { getCategoryList } from '@/shared/api/showcase';
+import { getCategoryList } from '@/shared/api/teams';
 
 const useCategoryListQuery = (): UseQueryResult<string[]> => {
   return useQuery<string[]>({
