+++ src/shared/hook/useClubListQuery.ts
@@ -1,7 +1,7 @@
 import { UseQueryResult, useQuery } from '@tanstack/react-query';
 
-import { getClubInfo } from '@/shared/api/showcase';
-import { ClubType } from '@/shared/api/showcase/type';
+import { getClubInfo } from '@/shared/api/teams';
+import { ClubType } from '@/shared/api/teams/type';
 
 export const useClubListQuery = (): UseQueryResult<ClubType[], Error> => {
   return useQuery<ClubType[], Error>({
